extract exporterConstants file
