+++ apps/src/util/exporterConstants.js
@@ -1,0 +1,11 @@
+/**
+ * @fileoverview Constants used by applab and gamelab exporters. Keep these
+ * separate so that they can be referenced without pulling in the exporter code
+ * and its dependencies (167KB).
+ */
+
+export const EXPO_SDK_VERSION = '34.0.0';
+
+export const PLATFORM_ANDROID = 'android';
+export const PLATFORM_IOS = 'ios';
+export const DEFAULT_PLATFORM = PLATFORM_ANDROID;
